refactor: move report printer to core module
Report printer a core function
